+++ src/main/java/generiek/api/EnrollmentEndpoint.java
@@ -299,6 +299,13 @@ public class EnrollmentEndpoint {
                     "Error in obtaining new accessToken with saved refreshToken for enrolment request:" + enrollmentRequest, e);
         }
         String accessToken = (String) oidcResponse.get("access_token");
+        String refreshToken = (String) oidcResponse.get("refresh_token");
+
+        //If all went well, save the new access token and refresh token in the enrollment request
+        enrollmentRequest.setAccessToken(accessToken);
+        enrollmentRequest.setRefreshToken(refreshToken);
+        enrollmentRepository.save(enrollmentRequest);
+
         String resultsURI;
         try {
             resultsURI = serviceRegistry.resultsURI(enrollmentRequest);